doorOpen + helper functions
created a function that opens the "door"/gate/cover. made a function that resets the lcd and door to the original state and a function that prints to the lcd when a user inputs an incorrect password. also changed the validate input function to be inside an if(key) condition.