Updated changePassword and validateInput functions
for both functions, made it so that if a user inputted more numbers than is expected for the password, even if they entered the password correctly, if they enter more numbers it's considered an invalid attempt. when creating a new password, the changePassword function ensures that a user entered enough inputs for a valid password. if they enter too little or too many numbers, they will be informed and the password will not be reset. it also checks to see if the "new" password they entered is different from the original one, if it's the same the user will be notified that the password didn't change.